Removed sync APIs around rediscovery
Async APIs are used all the way to the top public layer where blocking occurs.
