+++ driver/src/main/java/org/neo4j/driver/internal/cluster/RoutingProcedureClusterCompositionProvider.java
@@ -22,7 +22,6 @@ import java.util.List;
 import java.util.concurrent.CompletionStage;
 
 import org.neo4j.driver.internal.async.AsyncConnection;
-import org.neo4j.driver.internal.spi.Connection;
 import org.neo4j.driver.internal.util.Clock;
 import org.neo4j.driver.v1.Logger;
 import org.neo4j.driver.v1.Record;
@@ -51,13 +50,6 @@ public class RoutingProcedureClusterCompositionProvider implements ClusterCompos
         this.clock = clock;
         this.log = log;
         this.routingProcedureRunner = routingProcedureRunner;
-    }
-
-    @Override
-    public ClusterCompositionResponse getClusterComposition( Connection connection )
-    {
-        RoutingProcedureResponse response = routingProcedureRunner.run( connection );
-        return processRoutingResponse( response );
     }
 
     @Override
