+++ driver/src/main/java/org/neo4j/driver/internal/cluster/loadbalancing/LeastConnectedLoadBalancingStrategy.java
@@ -18,11 +18,8 @@
  */
 package org.neo4j.driver.internal.cluster.loadbalancing;
 
-import java.util.function.Function;
-
-import org.neo4j.driver.internal.async.pool.AsyncConnectionPool;
 import org.neo4j.driver.internal.async.BoltServerAddress;
-import org.neo4j.driver.internal.spi.ConnectionPool;
+import org.neo4j.driver.internal.async.pool.AsyncConnectionPool;
 import org.neo4j.driver.v1.Logger;
 import org.neo4j.driver.v1.Logging;
 
@@ -38,45 +35,29 @@ public class LeastConnectedLoadBalancingStrategy implements LoadBalancingStrateg
     private final RoundRobinArrayIndex readersIndex = new RoundRobinArrayIndex();
     private final RoundRobinArrayIndex writersIndex = new RoundRobinArrayIndex();
 
-    private final ConnectionPool connectionPool;
-    private final AsyncConnectionPool asyncConnectionPool;
+    private final AsyncConnectionPool connectionPool;
     private final Logger log;
 
-    public LeastConnectedLoadBalancingStrategy( ConnectionPool connectionPool, AsyncConnectionPool asyncConnectionPool,
-            Logging logging )
+    public LeastConnectedLoadBalancingStrategy( AsyncConnectionPool connectionPool, Logging logging )
     {
         this.connectionPool = connectionPool;
-        this.asyncConnectionPool = asyncConnectionPool;
         this.log = logging.getLog( LOGGER_NAME );
     }
 
     @Override
     public BoltServerAddress selectReader( BoltServerAddress[] knownReaders )
     {
-        return select( knownReaders, readersIndex, "reader", connectionPool::activeConnections );
-    }
-
-    @Override
-    public BoltServerAddress selectReaderAsync( BoltServerAddress[] knownReaders )
-    {
-        return select( knownReaders, readersIndex, "reader", asyncConnectionPool::activeConnections );
+        return select( knownReaders, readersIndex, "reader" );
     }
 
     @Override
     public BoltServerAddress selectWriter( BoltServerAddress[] knownWriters )
     {
-        return select( knownWriters, writersIndex, "writer", connectionPool::activeConnections );
+        return select( knownWriters, writersIndex, "writer" );
     }
 
-    @Override
-    public BoltServerAddress selectWriterAsync( BoltServerAddress[] knownWriters )
-    {
-        return select( knownWriters, writersIndex, "writer", asyncConnectionPool::activeConnections );
-    }
-
-    // todo: remove Function from params when only async is supported
     private BoltServerAddress select( BoltServerAddress[] addresses, RoundRobinArrayIndex addressesIndex,
-            String addressType, Function<BoltServerAddress,Integer> activeConnectionFunction )
+            String addressType )
     {
         int size = addresses.length;
         if ( size == 0 )
@@ -96,7 +77,7 @@ public class LeastConnectedLoadBalancingStrategy implements LoadBalancingStrateg
         do
         {
             BoltServerAddress address = addresses[index];
-            int activeConnections = activeConnectionFunction.apply( address );
+            int activeConnections = connectionPool.activeConnections( address );
 
             if ( activeConnections < leastActiveConnections )
             {
