+++ driver/src/test/java/org/neo4j/driver/internal/cluster/RoutingProcedureRunnerTest.java
@@ -26,8 +26,6 @@ import java.util.concurrent.CompletionStage;
 
 import org.neo4j.driver.internal.async.AsyncConnection;
 import org.neo4j.driver.internal.async.BoltServerAddress;
-import org.neo4j.driver.internal.spi.Connection;
-import org.neo4j.driver.internal.summary.InternalServerInfo;
 import org.neo4j.driver.v1.Record;
 import org.neo4j.driver.v1.Statement;
 import org.neo4j.driver.v1.Value;
@@ -37,10 +35,8 @@ import static java.util.Arrays.asList;
 import static java.util.Collections.EMPTY_MAP;
 import static java.util.Collections.singletonList;
 import static java.util.concurrent.CompletableFuture.completedFuture;
-import static org.hamcrest.core.IsEqual.equalTo;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
@@ -56,23 +52,7 @@ import static org.neo4j.driver.v1.Values.parameters;
 public class RoutingProcedureRunnerTest
 {
     @Test
-    public void shouldCallGetRoutingTableWithEmptyMap() throws Throwable
-    {
-        // Given
-        RoutingProcedureRunner runner = new TestRoutingProcedureRunner( RoutingContext.EMPTY );
-        Connection mock = mock( Connection.class );
-        when( mock.server() ).thenReturn(
-                new InternalServerInfo( new BoltServerAddress( "123:45" ), "Neo4j/3.2.1" ) );
-        // When
-        RoutingProcedureResponse response = runner.run( mock );
-
-        // Then
-        assertThat( response.procedure(), equalTo(
-                new Statement( "CALL " + GET_ROUTING_TABLE, parameters( GET_ROUTING_TABLE_PARAM, EMPTY_MAP ) ) ) );
-    }
-
-    @Test
-    public void shouldCallGetRoutingTableWithEmptyMapAsync()
+    public void shouldCallGetRoutingTableWithEmptyMap()
     {
         RoutingProcedureRunner runner = new TestRoutingProcedureRunner( RoutingContext.EMPTY,
                 completedFuture( asList( mock( Record.class ), mock( Record.class ) ) ) );
@@ -86,26 +66,7 @@ public class RoutingProcedureRunnerTest
     }
 
     @Test
-    public void shouldCallGetRoutingTableWithParam() throws Throwable
-    {
-        // Given
-        URI uri = URI.create( "bolt+routing://localhost/?key1=value1&key2=value2" );
-        RoutingContext context = new RoutingContext( uri );
-        RoutingProcedureRunner runner = new TestRoutingProcedureRunner( context );
-        Connection mock = mock( Connection.class );
-        when( mock.server() ).thenReturn(
-                new InternalServerInfo( new BoltServerAddress( "123:45" ), "Neo4j/3.2.1" ) );
-        // When
-        RoutingProcedureResponse response = runner.run( mock );
-
-        // Then
-        Value expectedParams = parameters( GET_ROUTING_TABLE_PARAM, context.asMap() );
-        assertThat( response.procedure(), equalTo(
-                new Statement( "CALL " + GET_ROUTING_TABLE, expectedParams ) ) );
-    }
-
-    @Test
-    public void shouldCallGetRoutingTableWithParamAsync()
+    public void shouldCallGetRoutingTableWithParam()
     {
         URI uri = URI.create( "bolt+routing://localhost/?key1=value1&key2=value2" );
         RoutingContext context = new RoutingContext( uri );
@@ -122,25 +83,7 @@ public class RoutingProcedureRunnerTest
     }
 
     @Test
-    public void shouldCallGetServers() throws Throwable
-    {
-        // Given
-        URI uri = URI.create( "bolt+routing://localhost/?key1=value1&key2=value2" );
-        RoutingContext context = new RoutingContext( uri );
-        RoutingProcedureRunner runner = new TestRoutingProcedureRunner( context );
-        Connection mock = mock( Connection.class );
-        when( mock.server() ).thenReturn(
-                new InternalServerInfo( new BoltServerAddress( "123:45" ), "Neo4j/3.1.8" ) );
-        // When
-        RoutingProcedureResponse response = runner.run( mock );
-
-        // Then
-        assertThat( response.procedure(), equalTo(
-                new Statement( "CALL " + GET_SERVERS ) ) );
-    }
-
-    @Test
-    public void shouldCallGetServersAsync()
+    public void shouldCallGetServers()
     {
         URI uri = URI.create( "bolt+routing://localhost/?key1=value1&key2=value2" );
         RoutingContext context = new RoutingContext( uri );
@@ -222,13 +165,6 @@ public class RoutingProcedureRunnerTest
         {
             super( context );
             this.runProcedureResult = runProcedureResult;
-        }
-
-        @Override
-        List<Record> runProcedure( Connection connection, Statement procedure )
-        {
-            // I do not want any network traffic
-            return null;
         }
 
         @Override
